Добавляет board-presenter
