+++ src/render.js
@@ -16,4 +16,5 @@ const render = (component, container, place = RenderPosition.BEFOREEND) => {
   container.insertAdjacentElement(place, component.getElement());
 };
 
+
 export {RenderPosition, createElement, render};
