+++ src/view/edit-point.js
@@ -1,6 +1,6 @@
 import {createElement} from "../render";
 
-const createEditPointTemplate = ()=>` <form class="event event--edit" action="#" method="post">
+const createEditPointTemplate = () => `  <li class="trip-events__item"> <form class="event event--edit" action="#" method="post">
                 <header class="event__header">
                   <div class="event__type-wrapper">
                     <label class="event__type  event__type-btn" for="event-type-toggle-1">
@@ -152,9 +152,9 @@ const createEditPointTemplate = ()=>` <form class="event event--edit" action="#"
                     <p class="event__destination-description">Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps, it's renowned for its skiing.</p>
                   </section>
                 </section>
-              </form>`
+              </form></li>`
 
-export default class EditPoint{
+export default class EditPoint {
   getTemplate() {
     return createEditPointTemplate()
   }
