+++ src/view/new-point.js
@@ -1,6 +1,6 @@
 import {createElement} from "../render";
 
-const createNewPointTemplate = ()=> `  <form class="event event--edit" action="#" method="post">
+const createNewPointTemplate = () => `   <li class="trip-events__item"> <form class="event event--edit" action="#" method="post">
                 <header class="event__header">
                   <div class="event__type-wrapper">
                     <label class="event__type  event__type-btn" for="event-type-toggle-1">
@@ -159,7 +159,7 @@ const createNewPointTemplate = ()=> `  <form class="event event--edit" action="#
                     </div>
                   </section>
                 </section>
-              </form>`
+              </form></li>`
 
 export default class NewPoint {
   getTemplate() {
